requirements/framework.txt
albumentations
beautifulsoup4
bezier
einops
modelscope
ms-swift
numpy
open_clip_torch
opencv-python
opencv_transforms>=0.0.6
oss2>=2.15.0
pycocotools
pyyaml>=5.3.1
scikit-image
torchsde
transformers
scikit-learn